very very weird bug fixed
when i opened the project today and starting added some new sprites in the the player started spawning itself on the root note instead of where i placed it. i couldnt find any fix this to this and i tried everything, so i ended up just remaking the player node and reimporting all the nodes and scripts and etc. it was so tedious and really frustrating but it works (probably)