finished designing level 3, playtested a LOT, altered difficulty, added player healing, added a count to how many enemies killed and how many needed total
i will add these new features to the mini tutorial at some point
the major problem here was def difficulty and the design of the level itself. basically, a lot of the replaying and retesting came from platforms being in the wrong places and spots being unreachable or hard to get to, especially with a level as big as level 3 is. it was also terribly difficult which is what led me to add player healing and edit a lot of the enemy stats. also, i had a problem with the level naming system so i just ended up changing it to file paths instead of getting the scene name to manage the scenes. decor is NOT done yet but i will add that too!!